requirements.txt
Click==7.0
DateTime==4.3
Flask==1.0.2
Jinja2==2.10
MarkupSafe==1.1.0
PyYAML==3.13
Werkzeug==0.14.1
certifi==2018.11.29
chardet==3.0.4
idna==2.7
itsdangerous==1.1.0
numpy==1.16.0
pip==19.0.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pymongo==3.7.2
python-dateutil==2.7.5
pytz==2018.9
requests==2.20.1
ruamel.appconfig==0.5.4	
ruamel.std.argparse==0.8.1	
setuptools==39.1.0	
six==1.12.0	
urllib3==1.24.1	
xmltodict==0.11.0	
zope.interface==4.6.0	 
